refactor: create new runner for sending release tweets and update script
